browsing/browser_initializer.py: wait for the cookies bar to disappear
If you interact too quickly with the page after clicking the "accept"
button the whole cookies bar might not close so we have to check if
it's gone. It never happened with firefox but occasionally happens with
chrome. This leads me to think it's a timing issue that might affect
slower machines/connections in any browser. Annyway, we'll always wait
for the element to be gone before continuing.